fix(ltx_2/tpu): host pipeline in fp32 — bf16 is emulated on the host (audio VAE 114 s, vocoder 196 s)
The first full LTX-2 request on the v5e took 450 s: the diffusers host
pipeline had been loaded in the DiT's bf16, and bf16 matmuls/convs are
emulated on the EPYC host. Measured in the worker with the runner's decode
split: audio VAE decode 114 s and vocoder 196 s in bf16 against 0.1 s and
0.6 s in fp32; the bf16 connectors were another ~40 s per prompt
(e2e-to-latents 93.6 s -> 50.7 s). The host pipeline (connectors, audio VAE,
vocoder; Gemma already was) now loads fp32, the prompt broadcast stays fp32
(the fp32 connectors consume it; the orchestrator casts to the DiT dtype
when it builds the bundle), and the video VAE goes to the chip in the DiT
dtype. The runner splits the decode timing per component and writes
BFCHW frames correctly.

Measured now (480x704x49, 20 steps, guidance 1.0): 49-51 s to latents
(1.45 s/step x 20 = 29 s DiT + Gemma-3 fp32 encode), decode 52 s of which
52 s is the video VAE's first compile (0.2 s warm), audio 0.7 s.

+++ tests/unit/models/ltx_2/test_ltx2_tpu_application.py
@@ -83,9 +83,11 @@ def test_broadcast_prompt_encoder_matches_diffusers_call_surface(monkeypatch):
         return torch.ones(1, max_sequence_length, 6, dtype=dtype), torch.ones(1, max_sequence_length)
 
     pipe = types.SimpleNamespace(_get_gemma_prompt_embeds=original)
-    mod._install_broadcast_prompt_encoder(pipe, is_encoder=True, dtype=torch.bfloat16, packed_width=6)
+    mod._install_broadcast_prompt_encoder(pipe, is_encoder=True, dtype=torch.float32, packed_width=6)
     embeds, mask = pipe._get_gemma_prompt_embeds("a fox", num_videos_per_prompt=1, max_sequence_length=8,
-                                                 scale_factor=8, device="cpu", dtype=torch.float32)
-    assert calls["prompt"] == "a fox" and calls["L"] == 8 and calls["dtype"] is torch.bfloat16
-    assert embeds.shape == (1, 8, 6) and embeds.dtype is torch.float32  # cast to the caller's dtype
+                                                 scale_factor=8, device="cpu", dtype=torch.bfloat16)
+    assert calls["prompt"] == "a fox" and calls["L"] == 8 and calls["dtype"] is torch.float32
+    # Stays in the host pipeline's dtype: the fp32 connectors consume it next;
+    # the orchestrator casts to the DiT dtype when it builds the bundle.
+    assert embeds.shape == (1, 8, 6) and embeds.dtype is torch.float32
     assert mask.dtype is torch.int64 and mask.shape == (1, 8)
